# authority/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Authority API
    version: 0.0.1
paths:
    /authority/v1/menu/create:
        post:
            tags:
                - Authority
            operationId: Authority_CreateMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/authority.v1.CreateMenuReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/authority.v1.CreateMenuResp'
    /authority/v1/menu/delete:
        delete:
            tags:
                - Authority
            operationId: Authority_DeleteMenu
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/authority.v1.DeleteMenuResp'
    /authority/v1/menu/detail:
        post:
            tags:
                - Authority
            operationId: Authority_DetailMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/authority.v1.DetailMenuReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/authority.v1.DetailMenuResp'
    /authority/v1/menu/list:
        post:
            tags:
                - Authority
            description: 菜单
            operationId: Authority_MenuList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/authority.v1.MenuListReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/authority.v1.MenuListResp'
    /authority/v1/menu/update:
        put:
            tags:
                - Authority
            operationId: Authority_UpdateMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/authority.v1.UpdateMenuReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/authority.v1.UpdateMenuResp'
components:
    schemas:
        authority.v1.CreateMenuReq:
            type: object
            properties:
                parentId:
                    type: integer
                    format: int64
                title:
                    type: string
                icon:
                    type: string
                path:
                    type: string
                permission:
                    type: string
                type:
                    type: integer
                    format: int32
                status:
                    type: integer
                    format: int32
                hide:
                    type: integer
                    format: int32
                note:
                    type: string
                sort:
                    type: integer
                    format: int32
                checkedList:
                    type: array
                    items:
                        type: integer
                        format: int64
        authority.v1.CreateMenuResp:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
        authority.v1.DeleteMenuResp:
            type: object
            properties: {}
        authority.v1.DetailMenuReq:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
        authority.v1.DetailMenuResp:
            type: object
            properties:
                info:
                    $ref: '#/components/schemas/authority.v1.MenuInfo'
        authority.v1.MenuInfo:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                parentId:
                    type: integer
                    format: int64
                title:
                    type: string
                icon:
                    type: string
                path:
                    type: string
                permission:
                    type: string
                type:
                    type: integer
                    format: int32
                method:
                    type: string
                status:
                    type: integer
                    format: int32
                hide:
                    type: integer
                    format: int32
                note:
                    type: string
                sort:
                    type: integer
                    format: int32
                createUser:
                    type: integer
                    format: int64
                createTime:
                    type: integer
                    format: int64
                updateUser:
                    type: integer
                    format: int64
                updateTime:
                    type: integer
                    format: int64
        authority.v1.MenuListReq:
            type: object
            properties:
                title:
                    type: string
        authority.v1.MenuListResp:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/authority.v1.MenuInfo'
        authority.v1.UpdateMenuReq:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                parentId:
                    type: integer
                    format: int64
                title:
                    type: string
                icon:
                    type: string
                path:
                    type: string
                permission:
                    type: string
                type:
                    type: integer
                    format: int32
                status:
                    type: integer
                    format: int32
                hide:
                    type: integer
                    format: int32
                note:
                    type: string
                sort:
                    type: integer
                    format: int32
                checkedList:
                    type: array
                    items:
                        type: integer
                        format: int64
        authority.v1.UpdateMenuResp:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
tags:
    - name: Authority